Reorganize color spaces into folders
