+++ Sources/ColorTokensKit/ColorSpace/LCH/LCHColor+Interpolation.swift
@@ -1,0 +1,21 @@
+//
+//  LCHColor+Interpolation.swift
+//  ColorTokensKit
+//
+//  Created by Siddhant Mehta on 2024-07-29.
+//
+
+import Foundation
+
+public extension LCHColor {
+    /// Interpolates between two LCH colors
+    func lerp(_ other: LCHColor, t: CGFloat) -> LCHColor {
+        let angle = (((((other.h - h).truncatingRemainder(dividingBy: 360)) + 540).truncatingRemainder(dividingBy: 360)) - 180) * t
+        return LCHColor(
+            l: l + (other.l - l) * t,
+            c: c + (other.c - c) * t,
+            h: (h + angle + 360).truncatingRemainder(dividingBy: 360),
+            alpha: alpha + (other.alpha - alpha) * t
+        )
+    }
+}
