+++ Sources/ColorTokensKit/ColorSpace/LCH/LCHColor+Stops.swift
