+++ Sources/ColorTokensKit/ColorSpace/RGB/RGBColor.swift
@@ -1,7 +1,6 @@
 //
 //  ColorRGB.swift
 //
-//
 //  Original repo: https://github.com/timrwood/ColorSpaces
 //
