+++ Sources/ColorTokensKit/ColorSpace/XYZ/XYZColor.swift
@@ -1,8 +1,7 @@
 //
 //  ColorXYZ.swift
 //
-//
-//  Built off of the original repo of: https://github.com/timrwood/ColorSpaces
+//  Original repo: https://github.com/timrwood/ColorSpaces
 //
 
 import Foundation
