+++ Sources/ColorTokensKit/Platform/AppKit/AppKit+Color+Ext.swift
@@ -1,0 +1,15 @@
+//
+//  Color+Ext.swift
+//  ColorTokensKit
+//
+//  Created by Siddhant Mehta on 2025-02-22.
+//
+
+#if canImport(AppKit)
+
+import AppKit
+/// Extension to `Color` providing support for dynamic colors based on the user interface style (light or dark mode).
+/// Source: https://www.jessesquires.com/blog/2023/07/11/creating-dynamic-colors-in-swiftui/
+/// Need to follow this source to build more
+
+#endif
